Add class for Ronchi test analysis

+++ rich/prototype/CbmRichProtRingFinderHoughImpl.h
@@ -76,7 +76,7 @@ protected:
 	vector<CbmRichHoughHit> fData; // Rich hits
 	vector<unsigned short> fHist; // XY histogram
 	vector<unsigned short> fHistR; // Radius histogram
-	vector< vector<unsigned short> > fHitInd; // store hit indexes for different group of hits
+	vector< vector<unsigned int> > fHitInd; // store hit indexes for different group of hits
 	vector<CbmRichRingLight*> fFoundRings; // collect found rings
 	CbmRichRingFitterCOP* fFitCOP; // COP ring fitter
 	CbmRichRingSelectAnn* fANNSelect; // ANN selection criteria
@@ -136,8 +136,8 @@ public:
 	 * \param[in] indmax Maximum index of the hit in local area.
 	 */
 	virtual void HoughTransform(
-	      unsigned short indmin,
-			unsigned short indmax);
+	      unsigned int indmin,
+			unsigned int indmax);
 
 	/**
 	 * \brief Main procedure for Hough Transform.
@@ -146,8 +146,8 @@ public:
     * \param[in] iPart Index of the hit group.
     */
 	virtual void HoughTransformGroup(
-	      unsigned short indmin,
-			unsigned short indmax,
+	      unsigned int indmin,
+			unsigned int indmax,
 			int iPart);
 
 	/**
@@ -178,7 +178,7 @@ public:
     * \param[in] hitInd Index in TClonesArray.
     */
    int GetHitIndex(
-         unsigned short hitInd);
+         unsigned int hitInd);
 
    /**
     * \brief Set fIsUsed flag to true for hits attached to the ring.
